2/6/2020

+++ Modules/Admin/Http/Controllers/AdminCategoryController.php
@@ -11,19 +11,24 @@ use Illuminate\Routing\Controller;
 
 class AdminCategoryController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     * @return Response
-     */
-    public function index()
+    public function index(Request $request)
     {
-        $data['listCategory'] = Category::select('cate_id','cate_name','cate_title_seo','cate_active')->get();
-        return view('admin::category.index',$data);
+        $category = Category::select('cate_id', 'cate_name', 'cate_title_seo', 'cate_active');
+        if($request->name){
+            $result = $request->name;
+            $result = str_replace('','%',$result);
+            $category = $category->where('cate_name','like','%'.$result.'%');
+        }
+
+        $data['listCategory'] = $category->get();
+        return view('admin::category.index', $data);
     }
-    public function getAdd(){
+    public function getAdd()
+    {
         return view('admin::category.add');
     }
-    public function postAdd(AddCateRequest $request){
+    public function postAdd(AddCateRequest $request)
+    {
         $category = new Category();
         $category->cate_name = $request->cate_name;
         $category->cate_slug = str_slug($request->cate_name);
@@ -33,14 +38,16 @@ class AdminCategoryController extends Controller
         $category->cate_active = $request->cate_active;
 
         $category->save();
-        return redirect()->back()->with('success','Thêm thành công');
+        return redirect()->back()->with('success', 'Thêm thành công');
     }
 
-    public function getEdit($id){
+    public function getEdit($id)
+    {
         $category = Category::find($id);
-        return view('admin::category.edit',compact('category'));
+        return view('admin::category.edit', compact('category'));
     }
-    public function postEdit(EditCateRequest $request,$id){
+    public function postEdit(EditCateRequest $request, $id)
+    {
         $category = Category::find($id);
         $category->cate_name = $request->cate_name;
         $category->cate_slug = str_slug($request->cate_name);
@@ -50,15 +57,24 @@ class AdminCategoryController extends Controller
         $category->cate_active = $request->cate_active;
 
         $category->save();
-        return redirect('admin/category')->with('success','Sửa danh mục thành công');
+        return redirect('admin/category')->with('success', 'Sửa danh mục thành công');
     }
 
-    public function getDelete($id){
-        Category::destroy($id);
-        return back()->with('success','Xóa danh mục thành công');
+    public function getAction($action, $id)
+    {
+        if ($action) {
+            $category = Category::find($id);
+            switch ($action) {
+                case 'delete':
+                    Category::destroy($id);
+                    return back()->with('success', 'Xóa danh mục thành công');
+                    break;
+                case 'active':
+                    $category->cate_active = $category->cate_active ? 0 : 1;
+                    break;
+            }
+            $category->save();
+        }
+        return back();
     }
 }
-
-
-
-
